# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApplicationError": {
        "additionalProperties": false,
        "description": "Explain a failure and provide an opaque reference for local logs.",
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "diagnostic_reference": {
            "pattern": "^[a-f0-9]{32}$",
            "title": "Diagnostic Reference",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "diagnostic_reference"
        ],
        "title": "ApplicationError",
        "type": "object"
      },
...
      "ConfigurationProfile-Output": {
        "additionalProperties": false,
        "description": "Store the checkpoint frequency as a whole number of minutes in seconds.",
        "properties": {
          "checkpoint_interval_seconds": {
            "default": 300,
            "exclusiveMinimum": 0.0,
            "multipleOf": 60.0,
            "title": "Checkpoint Interval Seconds",
            "type": "integer"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "schema_version",
          "checkpoint_interval_seconds"
        ],
        "title": "ConfigurationProfile",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Keep all application errors in one predictable response envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApplicationError"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "InstalledModel": {
        "additionalProperties": false,
        "description": "Describe one installed encoder/decoder pair without exposing its path.",
        "properties": {
          "compatibility_identifier": {
            "pattern": "^dense_v1_[0-9a-f]{64}$",
            "title": "Compatibility Identifier",
            "type": "string"
          },
          "export_reference": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Export Reference",
            "type": "string"
          },
          "format_version": {
            "enum": [
              1,
              2
            ],
            "title": "Format Version",
            "type": "integer"
          },
          "installed_at": {
            "format": "date-time",
            "title": "Installed At",
            "type": "string"
          },
          "maximum_payload_bytes": {
            "const": 1024,
            "default": 1024,
            "title": "Maximum Payload Bytes",
            "type": "integer"
          },
          "maximum_side": {
            "const": 1024,
            "default": 1024,
            "title": "Maximum Side",
            "type": "integer"
          },
          "minimum_side": {
            "const": 512,
            "default": 512,
            "title": "Minimum Side",
            "type": "integer"
          },
          "model_identifier": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,63}$",
            "title": "Model Identifier",
            "type": "string"
          },
          "profile_identifier": {
            "const": "test_only_v1",
            "default": "test_only_v1",
            "title": "Profile Identifier",
            "type": "string"
          },
          "source_identifier": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Source Identifier",
            "type": "string"
          },
          "status": {
            "const": "experimental",
            "default": "experimental",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "model_identifier",
          "compatibility_identifier",
          "source_identifier",
          "profile_identifier",
          "status",
          "format_version",
          "minimum_side",
          "maximum_side",
          "maximum_payload_bytes",
          "installed_at",
          "export_reference"
        ],
        "title": "InstalledModel",
        "type": "object"
      },
      "JobList": {
        "additionalProperties": false,
        "description": "Return saved job snapshots in creation order.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobSnapshot"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "JobList",
        "type": "object"
      },
      "JobSnapshot": {
        "additionalProperties": false,
        "description": "Describe a saved job without executing it or exposing secret data.",
        "properties": {
          "available_actions": {
            "items": {
              "enum": [
                "pause",
                "resume",
                "stop",
                "cancel",
                "supply_missing_inputs"
              ],
              "type": "string"
            },
            "title": "Available Actions",
            "type": "array"
          },
          "configuration": {
            "$ref": "#/components/schemas/ConfigurationProfile-Output"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "estimated_seconds_remaining": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Seconds Remaining"
          },
          "experiment_identifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Identifier"
          },
          "frozen_settings": {
            "additionalProperties": true,
            "title": "Frozen Settings",
            "type": "object"
          },
          "job_identifier": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Job Identifier",
            "type": "string"
          },
          "latest_event_identifier": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Event Identifier"
          },
          "metrics": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Metrics",
            "type": "object"
          },
          "operation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation"
          },
          "phase": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Phase",
            "type": "string"
          },
          "progress": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "requested_action": {
            "anyOf": [
              {
                "enum": [
                  "pause",
                  "resume",
                  "stop",
                  "cancel",
                  "supply_missing_inputs"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Action"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "paused",
              "stopped",
              "completed",
              "cancelled",
              "failed",
              "interrupted",
              "needs_input"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_identifier",
          "status",
          "phase",
          "configuration",
          "available_actions",
          "created_at",
          "updated_at",
          "progress",
          "estimated_seconds_remaining",
          "operation",
          "experiment_identifier",
          "frozen_settings",
          "result",
          "error",
          "metrics",
          "requested_action",
          "latest_event_identifier"
        ],
        "title": "JobSnapshot",
        "type": "object"
      },
      "ModelInstallRequest": {
        "additionalProperties": false,
        "description": "Identify one explicit installation so that network retries are safe.",
        "properties": {
          "client_request_identifier": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.:-]*$",
            "title": "Client Request Identifier",
            "type": "string"
          },
          "export_reference": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Export Reference",
            "type": "string"
          }
        },
        "required": [
          "client_request_identifier",
          "export_reference"
        ],
        "title": "ModelInstallRequest",
        "type": "object"
      },
      "ModelList": {
        "additionalProperties": false,
        "description": "List explicitly installed experimental models.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InstalledModel"
            },
            "maxItems": 8,
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ModelList",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local experimental training and review. Inference requires a qualified model.",
    "title": "StegoLab",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "description": "Merge original saved records with durable workflow jobs.",
        "operationId": "list_jobs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Jobs"
      }
    },
...
    "/api/v1/models": {
      "get": {
        "description": "List explicitly installed experimental model pairs.",
        "operationId": "list_models",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Models"
      }
    },
    "/api/v1/models/install": {
      "post": {
        "description": "Verify a registered export pair and record it as an experimental model.",
        "operationId": "install_model",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelInstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstalledModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Install Model"
      }
    },
    "/api/v1/models/{model_identifier}": {
      "delete": {
        "description": "Forget an installed model unless an inference job still needs it.",
        "operationId": "remove_model",
        "parameters": [
          {
            "in": "path",
            "name": "model_identifier",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,63}$",
              "title": "Model Identifier",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Remove Model"
      }
    },
...
  }
}